Fix InputAdapters to getting found
Plugins were sorted by DataValue classes. DataCells can therefore not be
used directly as inputs for the maps. We need to instead check every
superclass of that DataCell we want to get the Adapters for, aswell, and
then save the result so that we don't need to do that a second time for
that class.

Signed-off-by: squareys <squareys@googlemail.com>