Fix tiny bug
Example: input type is "MyMultiOutputListener" (my subclass of
MultiOutputListener), then the check would suceed, since I can assign this
type to MultiOutputListener.

When I try to assign NodeModuleOutputChangedListener, it would throw an
IllegalArgumentException.

If we would just simply switch to testing whether the item type is
assignable from MultiOutputListener, it could also be Object, which would
also resolve Object inputs with our logic here, which is not intended.

A simple check for the class being exactly MultiOutputListener.class is
therefore suficient.

Signed-off-by: Squareys <squareys@googlemail.com>